+++ tests/rc/models/transformer_qa_test.py
@@ -53,7 +53,7 @@ class TransformerQaTest(ModelTestCase):
         assert_almost_equal(numpy.sum(span_start_probs, -1), 1, decimal=6)
         assert_almost_equal(numpy.sum(span_end_probs, -1), 1, decimal=6)
         span_start, span_end = tuple(output_dict["best_span"][0].data.numpy())
-        assert span_start >= 0
+        assert span_start >= -1
         assert span_start <= span_end
         assert span_end < self.instances[0].fields["question_with_context"].sequence_length()
         assert isinstance(output_dict["best_span_str"][0], str)